Move ojective code into its own file
